added seller withdrawal functionality

+++ LocalDropshipping.Web/Data/Entities/Withdrawals.cs
@@ -8,7 +8,7 @@ namespace LocalDropshipping.Web.Data.Entities
         [Key]
         public int WithdrawalId { get; set; }
         public string? UserEmail { get; set; }
-        public int AmountInPkr { get; set; }
+        public decimal AmountInPkr { get; set; }
         public PaymentStatus PaymentStatus { get; set; }
         public string? TransactionId { get; set; }
         public string? ProcessedBy { get; set; }
